requirements.txt
fastapi==0.95.0
pypdf==3.0.1
langchain==0.0.170
langchain_ollama==0.1.0
langchain_community==0.0.10
sentence_transformers==2.2.0
faiss-cpu==1.7.4
uvicorn==0.18.2
